+++ src/Skyrim/VMAD/Fragment/FragmentQUST.cpp
@@ -147,7 +147,7 @@ uint32_t FragmentQUST::GetSize() const
 void FragmentQUST::Write(FileWriter &writer) const
 {
     writer.record_write(&unk1, sizeof(unk1));
-    uint16_t count = fragments.size();
+    uint16_t count = static_cast<uint16_t>(fragments.size());
     writer.record_write(&count, sizeof(count));
     fileName.Write16(writer);
     for (uint16_t i = 0; i < count; ++i)
@@ -159,7 +159,7 @@ void FragmentQUST::Write(FileWriter &writer) const
         fragments[i]->scriptName.Write16(writer);
         fragments[i]->fragmentName.Write16(writer);
     }
-    count = aliases.size();
+    count = static_cast<uint16_t>(aliases.size());
     writer.record_write(&count, sizeof(count));
     for (uint16_t i = 0; i < count; ++i)
         aliases[i]->Write(writer);
